// frontend/src/api/generated/core.ts
export interface components {
    schemas: {
// ...
        UserResponse: {
            /** Format: date-time */
            createdAt: string;
            /** Format: uuid */
            id: string;
            /** Format: date-time */
            lastLoginAt: string | null;
            /** Format: date-time */
            lockedUntil: string | null;
            /** @enum {string} */
            role: "CUSTOMER" | "OPERATOR" | "AUDITOR" | "ADMIN";
            username: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}